Bug Id: 461, 538, 539
PointDao.getNextPointId() was not being used correctly - this method should be called each time a point id is needed, not called once and then increment the value to get more ids.

When copying a ccu, the new carrier route that is created was not getting the correct route id from the new route.

When copying a device, the name must be different than other devices of the same category, class and type.  Added '(copy)' to the end of the device name when copying.