Bug: 367
Added code so that two points on the same device within a given point type cannot have the same offset.  This code is in the wizard - when creating a new point for a device, you can only choose offsets that are not in use for the point type.  It is also in the point edit panel - you can not change the offset to something that another point of the same type is already using.