Bug Id: 625
Updated commander to handle DBChange messages more seemlessly.

When a device change comes in, only the device in question is updated in the tree instead of the whole tree being reloaded.  When a route change comes in, only the route in question is updated in the route combo box instead of the whole combo box being reloaded.