Bug Id: 787 The problem was that the manipulation of the service after the call to service.bat was making the service invalid. This manipulation was necessary to add the service to the YUKON GROUP so that the web app service is visible within TDC.
The call the installer was using is exactly identical to that of the notification service. For whatever reason, this same call fails on the new Tomcat5 service.

For now, I will be removing this call in the installer. However, we still need to find a way to make the installer join the YUKON GROUP.