Bug Id: 709
Hashed Password Support
Reworked authentication to support both hashed and plain-text passwords in addition to supporting Radius and creating a framework to support LDAP in the future. This change removed the ability to get or set the password from a [Lite]YukonUser object. This resulted in a large number of changes to existing code.

See Technical Release Note "Hashed Password Support" for more information.

Note: this isn't complete yet. Database Editor is missing a dialog to change passwords.