Bug Id: 642
This should be the last major Long Load Profile checkin! This adds some error handling so that requests that porter no longer knows about will be removed from the queue and processing of other requests can occur. When a request is removed in this fasion, its onFailure handler will be called which in the current usage will cause a failure email to be sent.