YUK-3821: Meter Read Widget
Second refactoring of the device definitioin and attribute services. This replaces DeviceBase with LiteYukonPAObject in most of the method calls. Removes AttributeSource. Adds code to LiteFactory to transfer the type when a lite is created. Adds a hashCode method to DeviceBase. Fixes all of the related tests (sort of, Jason has to finish).