YUK-4100: Rework navigation and menus for the old 410, the new CSR, and the web commander pages
The old "commander" page has been split from the "showcase" page. The old showcase pages are still available from the bottom of the new device detail CSR page and are presented as part of the same module. The commander page is also available from the CSR page or from its own selection page. It is presented as being in its own module.