YUK-4308
Removed double click warning. You can now check and un-check every box at will. The onclick added the command to a list, and the unchecking added it to the list again. So if you checked then unchecked the box you would still attempt to send the command when clicking execute. Changed so that it searched the list, and if it is already in there, it is removed instead of added again. The check boxes now work as a checkbox should.

Also changed the execute button. If clicked with making no changes now reads, "no change was made"