YUK-4307, YUK-4246, YUk-4288
Changed Dynamictag table so only the disabled tags will go in.

Extracted CapControlCache to be accessible anywhere in Yukon via YukonSpringHook.

Added the :V indicator for substations and feeders, and it will propegate down the list. Made it so the tag will have feederDisabled if it was disabled from the feeder, and subDisabled if it was executed from the sub.

Wrapped a get(0) in SeasonScheduleDaoImpl so on an updated database it will not exception, instead it will print the error of which capcontrol device is missing the season schedule.

Added the two new messages to the CBCClientConnection so it will communicate with updated servers.

And some other little thing I forgot about.