YUK-4280, YUK-4067, YUK-4311
Temp Move now displays in a heirachy for easier navigation. Instead of one click, now you select the feeder then hit the submit button. The submit button is at the top of the page next to the control order input to bring attention to the control order input.

After submitting, the page is forwarded to a confirmation page displaying where the capbank was moved.

(removed extra mappings from cbcclientconnection)