YUK-4406: Yukon Alarm Monitor fails to start - FileNotFoundException YUK-4391: TDC (and possibly other) Classpaths are all messed up
Yes, this is a big checkin so late in the game, but I think it is pretty stable.

The biggest change is a rework of how the beanRefContext.xml file is used. There are now a few additional beanRefContext.xml files. When Spring loads, it looks for all beanRefContext.xml files on the classpath that it can find. A parent/child relationship now exists between the various contexts so that there is no longer duplication between the web and client versions. To this end, there is now a "common" context that has what was common between the old client and web contexts. As it turns out, the client context is now just an alias for the common context. One side effect of this is that starting up tomcat will implicitly do a dependency check on the common/client context. This happens because stuff in the common context cannot see stuff in the web context, but stuff in the web context can see stuff in the client context. In this new world, billing and dbeditor are special cases and declare their own contexts. This is done via beanRefContext files in their respective projects and YukonSpringHook.setDefaultContext calls in their main methods.

The next biggest change is the gutting of the foreign project (all files have been moved to billing). The shell of the project still remains and the build scripts and installer can keep pretending it is real.

The other changes involve the scrubbing of any reference to the foreign project from various manifest and classpath files.