YUK-4649 Previous Usage Readings do not display all useful values
Changed the algorithm for displaying recent items in the previous usage drop down. To this end, RawPointHistoryDao was changed to accept a maxResults parameter and to be smart about reversing the list if the start/stop arguments were reversed. In addition a new ChartInterval was added called DAY_MIDNIGHT. For grouping purposes, DAY_MIDNIGHT, considers midnight to fall on the previous day. In other words, the interval begins one millisecond after midnight.