YUK-4321 New Scheduler Phase I (move in/move out, high bill and manually submit a request to a group
This is the main checkin of the new job scheduler code. There will be some more checkins to follow because this doesn't include the SQL and it isn't actually activated in any of the beanRef files, so it shouldn't start running on anyone's computers.

This adds the quartz jar. This is currently only used for its cron parsing abilities.