YUK-4320  Editor for processing devices in a meter group en mass
Added the following to DeviceGroup and updated UI to use:
isEditable() - can update name, change parent, delete group
isModifiable() - can add/remove sub groups and devices

Improved javascript and controller validation for device groups ui.

Added functionality to DeviceNameTag so that it uses paoname when the device is not a meter or the other formatting cannot be applied.