YUK-4698 Resolve startup circular reference with cache and the connection pool
Made a few more changes to ConnPool. It turns out most of the remaining try/catch blocks were not catching exceptions that could actually be thrown. So, they were removed.

In addition, the cap control connection was the only one that didn't just call connectWithoutWait and get on with its life. It seemed to be waiting around so it could send a command, but this command was already set as the registration message. So, the connect call was changed. This should shave 15 seconds off of the startup time when cap control isn't running.