YUK-4230 Click on Back button after deleting Orphaned CBC caused HTTP Status 404 error
YUK-4699 When Back button is clicked on the Search Results Page, its taking to the "Webpage has expired" page

Changed how the Navobject works with JSF. Upon leaving jsp we stop setting the navigation now, and rely on the bookmark stack. (This is what we have been doing).

The change is that I stopped updating the Navigation and only set the bookmarks when moving through jsf pages. This leaves the PreviousPage variable as what we will want to return to when we get back to jsp pages.

Updated results.jsp and tabledata.jsp to get the nav object to determine where the back button returns us to. This removes all 404 errors, and unexpected back button behavior refered to in the bugs.