YUK-4750 Inventory, Purchasing, Service Order, and Account Action Log Pages Are Always Blue
Fixed the purple.jsp template page to correctly include the CannonStyle.css (or whatever was specified in the role property). Fixed all of the other JSP files in this checkin to use the correct syntax for including their PurpleStyles.css file (this will ensure that it is overridable as well). Fixed OutputHeadContentTag to not include the same CSS file twice within one section. Added a id to the body tag in the purple.jsp, this will aid in overriding styles on pages that use the purple layout.