YUK-4780 Device Detail Cleanup
I've stripped away some styles from the device detail widgets. I'm not trying to say that this is better, but I wanted to pull off some of the "unique" styles that have been added to some of these JSPs. Over the next month I know we'll continue to tweak the L&F of these pages.