YUK-4780 Device Detail Cleanup
There was a vision when I created the titledContainer and the hideReveal tag, but this got a little muddled when the boxContainer came along. The boxContainer had the hide feature of the hideReveal tag and the container-with-a-title feature of the titledContainer. So, this is an attempt to unify everything. All three of these tags are really the same and should work the same as much as possible. To this end I've created an abstractContainer that takes a type attribute that can be used to specify which type of container you'd like. It my mind, it would be nice if abstractContainer was called titledContainer, but that ship has sailed. titledContainer has come to mean "that container with the rounded corners." Oh well.