YUK-4780 Device Detail Cleanup
Adds methods for getting Meter objects by DeviceGroup. This is much more efficient than looping over a list of YukonDevices and looking each one up.

Adds a method to return a Meter comparator that uses the configurable meter name.

Adds a mechanism for caching the formatted name within the Meter object itself (it isn't that expensive to generate, but when using it in a Comparator it could get a little out of hand).