YUK-4804 capcontrol editor "create Substation", "create Subbus", "create Feeder", "create Capbank" navigation is messed up.
The URL that we wanted to go back to was lost since we only store a 2-page history in CitNav. Added a method to store a page for later use, and changed the affected pages to use it.

----------

YUK-4230 Click on Back button after deleting Orphaned CBC caused HTTP Status 404 error

related to the YUK-4804

----

YUK-4731 Back from Search results page

related to the YUK-4804