YUK-4327 Ability to override default attribute settings
Because we aren't really following this, I backed out the code the Stacey checked in earlier. The general change to the location of the attributes will remain, but the other lookup classes are being removed.

I also made several changes to the schema file. In its previous form it was not a valid schema file and the XML didn't validate against it. Now the schema file is syntactically correct and can be used to validate the XML file. This is used on startup to validate the file (which is useful given that user generated files are now allowed).