YUK-5076 Change The FDR Importer to work with 3-phase
Added a line so the point name will take the remainder of the string, this will catch the _A _B for the three phase imports.

The point creation can now take a buffer or file name, this will allow the point importer to be called and passed in data rather than requiring a file read.

Tweaked the importer to not write a file to the hard drive and call the point creation with the new method.