YUK-5084 YUK-5078 Alter Schedule Options
Added two options to the schedule list. One for Time Sync and the other for verifying banks not controlled in a time frame.

Added an input field so the user can alter the text of the command going into the database, this will allow the time to be changed for the new command.

Removed the Info column. This column had no value other than a pop up with the text "Popup Text 1" hardcoded into it.