YUK-5202 Rewrite "Standard Page" implementation to better support i18n
(This checkin is closely tied to the YUK-4880 checkin that is going to follow it.)

A rewrite of the original standard page implementation. The original implementation was heavily based on several tags and some complex relationships between those tags. For the most part, those tags still exist, but now they serve mostly as data holders. The heavy lifting is now performed by the LayoutController class--a standard Spring MVC controller.

The menu render has also changed to be more i18n friendly. The module_config.xml no longer contains the actual strings that would be displayed on the menu. Instead, the keys serve this purpose now and are used to look up the actual display values in a properties file.