YUK-4880 Multi language option for end user pages in the DR market
This is the basic framework for support internationalization in Yukon. The biggest single change is the introduction of a YukonUserContext object. This object contains the locale, time zone, and the LiteYukonUser. Many of the services that used to accept a LiteYukonUser as one of their arguments have been changed to accept a YukonUserContext object. This allows dates (and points) to be consistently formatted correctly for the user. A YukonUserContext object is placed in all HttpServletRequests that come in and can be pulled out with the YukonUserContextUtils object similarly to how the LiteYukonUser could be retrieved with the ServletUtils class. Making this change touched dozens of files all over the code base.

The other change is the introduction of a cti:msg tag that can be used to pull messages out of one of the xml properties files (that live under common/i18n). This tag should be used to place text on a web page in an i18n safe manner.

To support the previous two changes, standard Spring MessageSource and Theme classes have been used to set the whole system up. Because of Yukon's "special" heritage, there are several customizations to get everything working, but the heart is all Spring.