YUK-5220 Fix LCUtils to correctly calculate and format dates
Modified LCUtils to be passed a YukonUserContext and use the DateFormattingService to format dates. I considered changing the get*ValueAt methods to simply return dates, but that would have made the "---" value tricky to handle.

getTimeString was modified to use the default system calendar to turn a number of seconds into a date because it should be considered time-zone-less. The user's context is then used to format that date into a string.