YUK-4972 add 3-phase point values to one-line view.
Added new Role property and added it to the one-line page.

YUK-5199 can't delete a substation object from the substations.jsp or feeders.jsp page.

Made changes to the cache to throw a NotFoundException when something is not in the maps. Also made changes to the affected files that relied on the null and changed them if they needed to catch the exception.