YUK-5221, YUK-4823, YUK-5025, YUK-5092
YUK-5221, YUK-4823 -- cursors now work in Oracle.  A lot of the original code has been removed because it ended up doing nothing.

YUK-5025
Added foreign key with error blocks.  Needed in 4.0 updates too in case the upgrade is from 3.5

YUK-5092
CCOPERATIONS_VIEW got in with error-clausing AS statements in the WHERE clause. Fixed.