YUK-4998, Module dropdown in the menu links header bar is not useful.
The drop down has been changed to use the links on the operations page.  This drop down also includes any checks of roles or properties.  I also had to include a way of passing in javascript files through this drop down, so that some drop down elements that need javascript, such as client launchers, could be used.  I also added the drop down labels to the menuLabels.xml to allow for easy internationalization.