YUK-5386 Create an web-based alert system for displaying alarms and other events to the user
This builds on the existing data updater functionality to constantly check if a new alert is present. If the number of alerts is greater than 0, the "Alerts" option is shown in the menu bar. Clicking this link will display the alerts popup window. This popup window is built on the newly expanded simplePopup tag that is also used for the JWS popup.

 - A long standing bug in the data updater was fixed that prevented the callback feature from working correctly.
 - Role properties have been added to control how long alerts are maintained and which login groups can see alarm alerts.
 - A new "common" servlet has been created to serve as a top level home for things that don't fit into "amr" or "dr".