YUK-5233, DBToolsframe is treating start-blocks like error-ignores YUK-5105, dbupdater starting at the beginning of a valids file instead of after last line processed before.
I had to pull the setIgnoreState method out of the for loop.  The check to see if it was previously successful was masking all of the ignore-end statements.  I also had to add a null case to the successful method to account for the false messages that were run while in the ignore state.  This took care of the problem of rerunning part of the script on start.