YUK-5442, Add a DEVICE_LABEL column to the STARS hardware import YUK-5507, StarsActionController brakes with any file input in stars. YUK-5502, Fix Deprecation Issue of DiskFileUpload in Stars.
Looks like most of the Device_Label add-on was already functional, but ran into a issue where the whole importer was not working.  I fixed the importer, and also fixed the StarsActionController issue that was related to the importer not working correctly.  I also pulled over the multipartResolver into STARS, which was used to fix the StarsAction issue.  As for the DiskFileUpload Deprecation, I have fixed a few instances of it, but not all of them yet.  Another addition in this commit is the addition of the commons-fileupload-1.2.1.jar.  This is the latest jar from apache that was needed to fix the deprecation issue.