YUK-5565 Add OPC import suport to the fdr importer
Altered ProgressImporter to be the fdr importer and interfaced the code that will change. Allowing for future fdr import styles to be added simply by creating and Parser for the text input.

Changed the old text import to use the new interface method.
Added an OPC parser.
Changed the main function to take an additional argument to choose the parser.