YUK-5440, Automate the stars customer/hardware import process and archive old import attempts.
Added an atomic boolean to allow other threads to check the status of database cache.  Added a new folder structure to go with the new archiving, failure logs, and scheduling that was added, which were also added in this commit.  Cleaned up some of the error handling and also converted to a run after fail instead of the fail on fault cause it used to have.