YUK-5634 Printable view zooms to size of the page. YUK-5633 Pop-ups do not work when zoomed in on One Line page.
The page no longer shrinks to fit everything on one screen. Large amounts of feeders will cause the browser to scroll. Zooming is no longer needed, and the page layout when printing will use the whole page.

YUK-5509 Trying to Perform a Temp Move Back from a One-Line Causes an Error to Occur
YUK-4603 Temp move on the one-line. and 4-tier (feeders.jsp)

Fixed the link to use the new controller for commands. Changed one-line to pop up a menu with the links to move banks or move back. This is more in line with the rest of one line, and allows the user to abort the move or move back. Screen shots attached.