YUK-5655 Lose breadcrumb for Special SubArea page after the feeders page is accessed
Special area Status was being lost when going to the feeders page. The page was making the determination off of the parent id. Which will always be an normal area.

Added a specialArea condition to the links from the substation.jsp page, and edited the feeders bread crumbs to have this flag as well.