YUK-5630, IP address resultion in DB into string, instead of decimal form YUK-5729, Delete unused TrendingRole properties YUK-5403, Is the status of whether a CB has been temp moved available in a DB view
Added all entries to the scripts.  I also had to create a special conflict script to the creation scripts folder.  This will allow us, for the short term, to add views or tables that we could not have added before b/c there was no joint solution for sql and oracle.