YUK-5862, STARS - Hardware Import Log Files Are Confusing YUK-5860, STARS - Import Account Data Feature Cannot Accept Files Larger than 100000 bytes
I made a small change to the file naming scheme to make the log files less confusing, and also modified modified the ActivityLog to not through null pointer exceptions.  As for YUK-5860, we pulled out the default values for upload max in a few files, amr-servlet.xml, deviceConfiguration-servlet.xml, group-servlet.xml, and stars-servlet.xml, and added it to applicationContext to centralize the value.  We also set the value to its default -1, which is no limit, to fix the 100000 cap issue we were running into.