YUK-5394 Database reports do not give collection group column YUK-4822 Database reporting doesn't show collection or alt group anymore. Added CollectionGroup and Test Collection Group to the following report models: - Carrier DB Model - Disconnect Collar Model - Meter Read Model (for Missed only) - Route DB Model (also added Billing Group).
Since meters are no longer limited to belonging to more than one /Meters/Collection/ group (for example), this report will output only 1 of the groups in the parent group if there are multiples.  When the group name is not unique, then an * will be placed at the end of the groupname.  The footer has a key for this.
Also changed "Disabled" to "Enabled" to closer follow naming convention used in bulk operations.