YUK-6041 Rewrite of Swing client login to better support JWS
Changes the login dialog to include a URL field instead of separate host and port fields (this also allows the clients to use https for the login check). Additionally, the JWS dialog now includes read-only fields and all dialogs have a select-on-tab behavior. Removes the idea of a session; all "End Session" menus are no gone. Changes all main methods to use a common ClientStartupHelper to ensure that they each do the exact same sequence for startup.