YUK-6181 Bulk Ops - devices counted more than once
Fixes getDeviceCount so that it uses the getDeviceGroupSqlWhereClause. Change several versions of the getDevices methods to return Sets instead of Lists (this fixes the root problem with the wrong count). Added a new base class for providers that implements four of the base methods on top of the where clause methods. Provided an overridden getDeviceGroupSqlWhereClause that has special handling for static groups to prevent ridiculously long SQL statements. Changed a tag to prevent duplicate calls to count (which unfortunately cannot be easily cached).