YUK-6276 Group Commander speedup followup
Removes command authorization checks because they were very slow and had little use since the group commander checks out going command before they are sent to the CommandRequestExecutor.

Changed the System Log code that was just checked in. It nows logs every command (instead of just device commands), does not call the PaoDao (which was the problem I was trying to solve above), and logs just after the command is actually sent to porter.