YUK-6269 DB Changes must be sent after a transaction commits
This is the main checkin for this bug.

Adds support for sending db change messages after a database transaction has committed. Spring Transaction support is used to essentially "synchronize" the sending of the db changes with the whole transaction. In addition, because the db changes are bunched together they can be reduced to remove duplicates and sequences that don't make sense (e.g. removing an update that follows an add).