YUK-6540 Clean up DBPersistent code
Cleans out unneeded or commented out code. Moves update method to CommonMulti so that it can be used by SmartMulti and the regular Multi. Replaces "don't call me" log statements with exceptions. Changes DBPersistentDao to simply execute each object in a transaction instead of creating a multi and executing that.