YUK-5719 ccu simulator - support multiple simultaneous connections
Hijacking this bug id. I altered the connections so that in the event of a crash, the thread will be restarted. Also changed so we will not get stuck in a loop if we get bad data from the buffer. The socket will restart if we get an invalid message.