YUK-6682 - Yukon Setup Shortcut on desktop is replacing the current IE window that is open, it should open up a new window.
The installer will now create two more shortcuts on the desktop.  These should always open in IE and always open in a new window.  The old shortcuts can still be used to open the yukon pages in the default browser.

Also modified the links in the Yukon Debug Servers folder so that the console will not close when the application closes or crashes.