YUK-6632 ISOC modifications for 2009 YUK-6724 Billing file containing event data to be transfered to PBS
Changes to Spring context hierarchy to support a dbeditor->billing->commercialcurtailment dependency. This throws out the old hibernate/non-hibernate separation which greatly simplifies the beanRefContext.xml files. The downside is that some of the clients may need more memory, but in my limited testing I couldn't prove or disprove this.