YUK-6856 TDC fails to render a display if existing rows exceed the tdc_max_rows role property
Ensures that pointValues is always adjusted to keep pointValues.size() and getRows().size in check. The createDummyPointValue method is a little strange because it uses getRowCount() to determine what to do. However, just before this method is called, a row was added, so the old code that would bail out if location == getRowCount() was bad because rows were always added regardless of the location.