YUK-6607 Cap Control magnifying glass view is very slow
Changed page to use the point formatting service. Once YUK-6606 is checked in the speed improvements from that will affect this page.

Made a change to The PointFormattingService, it was incorrectly caching states mapped to values causing a problem if more than one state group was represented on a page.