YUK-6606 Increase loading speed of Device Detail page
Adds a new feature to the DataUpdaterService where the backing services can indicate if they would return a value immediately for a given identifier. All existing backs will always indicate yes except for the point backing service. The point backing service will only return true if the point value is available in its cache. In operation, the DataUpdaterService will use ask if a response is available immediately when it is looking for the initial value (i.e. when the page is first rendering), on subsequent requests this check is skipped. This allows a page to render quickly with a placeholder value instead of having to wait for each point to come back from dispatch. When the page starts its background polling, the placeholder values will quickly be replaced with the actual values. When the DataUpdaterService is called from a background poll, it will happily wait forever to get any values.

As an additional optimization, a controller can hint both the point backing service and the point formatting service that various points are on the way. This allows the dispatch requests to be made ahead of time in bulk and allows the point formatting service to cache the LitePoints for the durration of that request.

The AsyncDynamicDataSourceImpl has been changed. There were some synchronizations problems that had gone unnoticed. Also the data structures have been changed to use Multimaps which greatly simplified the code that managed them.