YUK-6922 Clients may hang when starting up remotely with logging set to debug
The primary fix is simply ensuring that the HttpInvoker code does not log. Because we subclassed the invoker, its logging level was set with the com.cannontech loggers which made it hard to turn on debugging without breaking things. We no longer needed a custom version of the invoker, so it has been removed.