YUK-6599 - CapControl : After creating CapBank, Clicking on Return button ... YUK-5712 - In View, Strategies - clicking on Return...
Changed navigation to use 'GETS' instead of 'POSTS' so the browser back button can be used in place on the eye sore Back button in some JSP pages.

Restored faces navigation to work with the history stack, Faces pages will still need to use the return buttons but it will be isolated to the faces pages.

Made some style changes to a few cap control pages with the menu options and bread crumbs.