YUK-7001 Cannot edit User Roles for Login
Adds catch for UserNotInRoleException. While fixing this, found a strange problem with PK collisions when writing the user roles to the database. Switching to a sequence seemed to solve the problem, though. Also removed a method in the Point class I ran across that is no longer being used (and should not be called anyway).