YUK-6767 "Remember me" option on Yukon login page doesn't do that
The primary fix is URLEncoding the Base64 string that is assigned to the cookie. This is somewhat humorous because the cookie documentation states that Base64 would be a great option for putting binary data in a cookie, but it fails to mention that the equal sign that can appear in Base64 is an illegal cookie value.

While I was at it I changed the encryption to use a CBC mode cipher instead of the ECB. The ECB mode causes most of the encrypted output to look the same, which doesn't seem very secure.