YUK-7266 DB Editor is very slow when running against BGE sized database
I have focused on the following areas: 

The display of Logins at the bottom of the Login Group panel. 
The main tree of Logins. 
The main tree of Contacts. 
The main tree of CICustomers. 
The drop down of all Contacts in the Customer panel. 

When selecting System from the View menu, the default selection is now Login Groups. 

Along the way I've ripped out a couple large pieces of ServerDatabaseCache that are no longer needed, reorganized the ContactNotificationDao (including moving one method to it from ContactDao), and deprecated the getLiteContactNotifications() method on LiteContact (I removed all but a few references to it that I felt were too risky).

(all 4.2 checkins are rolled up into this one)