YUK-7209 WCTP is not working in head
Updated Xerces from 1.7.0 to 3.0.1.  This fixes the issues we had with linking with MSVCRT6D.DLL.  Put it in yukon-3rdparty where it belongs.

Changed the way that the Initialize() and Terminate() were called.  They were not being used quite right, which was resulting in exceptions being thrown from inside the library.