YUK-6785 Device Group escaping bug leads to security and usability issues
Changes the DeviceGroup Service and DAO methods to return the where clauses as SqlFragmentSources instead of Strings. This allows SQL parameters to be safely embedded in the where clause (as is required for the dynamic device type groups). This change required updating a lot of code that was building SQL statements with the String version of the where clauses. Any code that wasn't already using the JdbcTemplate required updating.

(This needs a bit more testing, particularly the billing code.)