YUK-7342 AMR event flags- tamper flag only YUK-7341 AMR Dashboard - Outage processing YUK-7356  Integration to RF vendor
Redo of the Service Manager to support services running in their own Spring context as well as services running on multiple computers.

Adds ActiveMQ as an embedded service that can be enabled with the YukonServices table.

Changes first checkin of the point processing code to support the new Service Manage code and support periodic refreshes (reloading the outage or tamper monitor data from the DB).