YUK-7970 - Moving the information icons out of the widget container tag
Removed the question mark icons from the outage monitors widget, the tamper flag monitors widget, and the scheduled reads widget.  Replaced them with a link at the top of the widget body, similar to the disconnect widget.

Also made some changes to the i18n keys.  It didn't seem like we needed html in the i18n strings.