YUK-5321 Automatic control with multiple programs/groups/gears is having strange text output
LM was goofing up the calculation of the control area state by only looking at the number of active programs of the current priority, instead of every program.  This caused the control area to be marked Inactive even if it had running programs.  The time control code then fixed this issue by calculating properly, but it output a Timed Start message to the log because it thought it had started an already running program.