YUK-7611 New, more generic, event logger YUK-8048 VEE advanced page for Accept All, Delete All, and go back 12 months
A commingled checkin of improvements to the event logger UI (now located in the Support module) and a helper/easy/advanced page (living in a new "monitors" module for now).

One major change worth noting: all of the event argument indexes have been increased by one. Only a programmer could appreciate that the argument columns were named starting with 1 and the i18n file referred to arguments starting with 0.

SqlStatementBuilder was changed to make better use of its own internal methods.

ServletUtil was updated with a helper for turning String[] into Enum collections.