YUK-8049 - Assigning schedules to a device causes a page blip in firefox
Set the main div of the page to not display on initial load.  After the appropriate rows are set, the div is then displayed.  
This still takes a fraction of a second to show the whole page, but it prevents the page from showing one set of results on load that is replaced by another set a second later.