YUK-8405 Webserver gets locked up when database goes offline and/or network interruptions occur
Removed the synchronization on the getGroup method because it was causing systems to lock up (doing database work inside a mutual exclusion lock, especially when inside of a method that could be transactional, is probably a bad idea). I don't think a case-insensitive cache is needed. We'll now just cache whatever we're asked for.