YUK-8010 It's possible to have a program in a scenario that's not in a control area
Moved dr.scenario from loadcontrol to common so it is accessible from database editor.

The Control Area Program Assignment panel no longer allows programs to be removed if they are assigned to a Scenario. Instead, when the user clicks the Remove button, a warning popup will appear, describing the specific problem and stating "Programs cannot be removed while assigned to a Scenario."

If several programs are selected from the list for removal, the warning message will list an error message for each conflicting program.