YUK-8756 Modify capcontrol table pages to look like new DR tables
Schedules, Strategies, Schedule Assignment and Recent Cap Bank Moves now use the compact results table, with paging. All the results are still loaded from the database, but with the limited number of results rendered on the page, these all load in a second or two on a Progress database. 

In addition, filtering on Schedules and/or Commands was added to the results box for the Schedule Assignment page.