YUK-8804 Yukon Installer 5.1+: Repair option stops services and does not restart them.
Installer now warns the user to stop services if any are running before a modify or repair install. If services are still running before the data transfer step, the installer attempts to stop them. After a modify or repair, if any services are installed, the user is given the option to start them.