YUK-8767 DeviceReadingsModel query doesn't correctly handle mapped attributes
Implemented a new method in the RawPointHistoryDao that returns the correct point data for an attribute.

Changed DeviceReadingsModel to use this new method in the RawPointHistoryDao instead of its own incorrect SQL.

Fixed a bug in the other RawPointHistoryDao methods where duplicates rows were being mistakenly returned.

A helper method was created in a new IterableUtils class for guessing the size of Iterables.

The ChunkingMappedSqlTemplate was changed to return a more specific Multimap (the actual object being returned was not changed, just its declared type).