YUK-9139 Setup.jsp page makes unintended database updates (Oracle) YUK-9142 Setup.jsp makes unintended database updates (SQL Server)
Changed RolePropertyEditorDaoImpl to save the actual default value instead of null (which would then be converted to an empty string). This has similar behavior but allows DB Editor to display the default value as it has done in the past. This change effectively addresses both issues. But to address the YUK-9139 directly, the value is passed through the standard SqlUtils methods to ensure that an empty String isn't persisted for Oracle's sake.